pyproject.toml
[project]
name = "guba-e-ethiopian-legal-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "groq>=1.1.2",
    "huggingface-hub>=1.8.0",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.4",
    "pinecone>=8.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "uvicorn>=0.42.0",
]
